[#4] refactor: core내 Event 도메인 core-infra 모듈로 분리
