+++ mega-coffee-core-infra/build.gradle.kts
@@ -1,0 +1,5 @@
+description = "mega-coffee-core-infra module"
+
+dependencies {
+
+}
