+++ mega-coffee-core-infra/src/main/kotlin/com/meloning/megaCoffee/core/infra/event/EventSender.kt
@@ -1,4 +1,4 @@
-package com.meloning.megaCoffee.core.event
+package com.meloning.megaCoffee.core.infra.event
 
 interface EventSender {
     fun send(type: EventType, payload: Map<String, String>)
