+++ settings.gradle.kts
@@ -24,6 +24,7 @@ include(
     "mega-coffee-api",
     "mega-coffee-common",
     "mega-coffee-core",
+    "mega-coffee-core-infra",
     "mega-coffee-infra:database:mysql",
     "mega-coffee-infra:message-queue:rabbitmq",
     "clients:java-email",
